feat: first real Polygon ingestion, end-to-end verified
Ingests AAPL for 2026-07-29 through the full path: Polygon API -> Parquet
archive -> raw.prices -> dbt staging view. Completes the Phase 1 checklist;
the 5 dbt tests now pass against real rows instead of vacuously on an empty
table.

Three issues surfaced only once real data flowed:

TLS interception. Avast Web/Mail Shield MITMs HTTPS with its own root CA,
which lives in the Windows trust store. Python verifies against the certifi
bundle instead, so every request failed with "unable to get local issuer
certificate". src/common/tls.py routes verification through the OS trust
store via truststore. Verification stays enabled - this changes which roots
are trusted, not whether trust is checked.

Fractional volume. Polygon reports volume as a float (v=56090840.685498,
fractional-share trades aggregated), which crashed the Int64 cast in
validate(). Rounding in the raw layer would violate "raw stays raw", so
raw.prices.volume becomes NUMERIC(20,6) holding the exact source value and
stg_polygon__prices does round(volume)::bigint. Regression test added.

Credential leak. requests embeds the full request URL in exception messages,
so an ?apiKey= query param reached the logs and was persisted to
pipeline_runs.error_message. The adapter now authenticates with an
Authorization: Bearer header. The affected ledger row was redacted in place.

Note: init.sql only runs on a fresh volume, so the volume column change was
applied to the running database with a manual ALTER TABLE (the dependent dbt
view had to be dropped and rebuilt). There is no migration tool yet.

+++ docker/postgres/init.sql
@@ -48,7 +48,10 @@ CREATE TABLE IF NOT EXISTS raw.prices (
     high            NUMERIC(18, 6),
     low             NUMERIC(18, 6),
     close           NUMERIC(18, 6),
-    volume          BIGINT,
+    -- NUMERIC, not BIGINT: vendors report fractional volume (Polygon aggregates
+    -- fractional-share trades, e.g. 56090840.685498). The raw layer stores what
+    -- the source sent; rounding to a whole share count happens in dbt staging.
+    volume          NUMERIC(20, 6),
     vwap            NUMERIC(18, 6),
     trade_count     INTEGER,
     source          VARCHAR(50)  NOT NULL,
